MindEase_Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic
requests
cryptography
python-dotenv
openai
